chore(all): leverage shared devdeps and scripts
affects: @rentpath/react-ui-ag, @rentpath/react-ui-core, @rentpath/react-ui-rent,
@rentpath/react-ui-rentals, @rentpath/react-ui-tracking, @rentpath/react-ui-utils

This commit removes the devDependencies from each package and merges them into
the devDependencies of the root package. Because of how Node handles module
resolution, we can leverage the shared node_modules directory of a parent dir
when looking for modules during code execution (in development). Any actual
dependencies still remain in their respective package.json files.

Additionally, this commit removes the scripts from each package and updates the
root package.json's scripts to leverage `lerna exec` instead of `lerna run`
which will execute the given command in each package, rather than looking for a
script of that name in each package as `lerna run` does. Because of this change,
we can further simplify each package by removing the `.babelrc` and the
jest configuration/setup from each package.

One of the biggest benefits to this code reorganization is that we can now
easily manage keeping our development dependencies up to date using Greenkeeper,
etc. as those services only work with a single package.json at the root of the
repo.

Furthermore, any changes to our test config, etc. will automatically be shared
across all our packages.

A common concern with these changes is that the tests are no longer isolated to
a single package, but in practice, the tests run quickly enough that this should
not be an issue. On my development machine, I see full runs of the test suite
regularly finish in under 10 seconds. Utilizing the --watch flag, etc. to limit
the tests to just the files changed will result in even faster times.