fix(Dropdown): check event target and event parent with specific data for handleDocumentClick
affects: @rentpath/react-ui-core

It is possible with event queueing that the caught event is that of the immediate child.
Include a
new and specific data attribute to reliably check to see if the Dropdown or the anchor is the target
of the click.
Also clean up setState logic in toggleVisibility to not rely on state before and
after setState call.