fix(photo): ssr missing img tag
affects: @rentpath/react-ui-core

Issue: We render the \`<img/>\` tag with SSR before. Bbut what happens is the DOM img is loaded in
concurrently while the client bundle is loaded in. So trying to use \`onError\` which react intercepts
ends up in a race condition where the img throws an error in the dom first potentially.
Fix:
Offload setting of src to the handler. Check for differences in src !== fallbackUrl and make
adjustments accordingly. State has to be fired if these 2 are the same so the mounted component will
update and force a re-render.